Issue #24858 move metadata out of persistence service package
