+++ dev/io.openliberty.data.internal.persistence/src/io/openliberty/data/internal/persistence/DataProvider.java
@@ -68,8 +68,8 @@ import io.openliberty.cdi.spi.CDIExtensionMetadata;
 import io.openliberty.checkpoint.spi.CheckpointPhase;
 import io.openliberty.data.internal.persistence.cdi.DataExtension;
 import io.openliberty.data.internal.persistence.cdi.FutureEMBuilder;
-import io.openliberty.data.internal.persistence.service.DataComponentMetaData;
-import io.openliberty.data.internal.persistence.service.DataModuleMetaData;
+import io.openliberty.data.internal.persistence.metadata.DataComponentMetaData;
+import io.openliberty.data.internal.persistence.metadata.DataModuleMetaData;
 import io.openliberty.data.internal.tracker.ModuleTracker;
 import io.openliberty.data.internal.version.DataVersionCompatibility;
 import jakarta.data.Limit;
