+++ dev/io.openliberty.data.internal.persistence/src/io/openliberty/data/internal/persistence/service/ResRefDelegator.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023 IBM Corporation and others.
+ * Copyright (c) 2023, 2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -22,12 +22,10 @@ import com.ibm.wsspi.resource.ResourceInfo;
 import io.openliberty.data.internal.persistence.DataProvider;
 
 /**
- * A resource factory that delegates to a JNDI name that was specified
- * as the dataStore for a Repository. The ResourceInfo is ignored because
- * the JNDI name is likely to already be a resource reference lookup
- * with its own ResourceInfo that should be used instead.
+ * A resource factory that delegates to a resource reference JNDI name that was
+ * specified as the dataStore for a Repository.
  */
-class DelegatingResourceFactory implements ResourceFactory {
+class ResRefDelegator implements ResourceFactory {
     private final String identifier;
     private final String jndiName;
     private final DataProvider provider;
@@ -35,17 +33,25 @@ class DelegatingResourceFactory implements ResourceFactory {
     /**
      * Construct a new instance.
      *
-     * @param jndiName JNDI name to look up obtaining resources.
-     * @param metadata metadata identifier, computed from the class loader identifier
-     *                     of the class loader of the repository interface.
+     * @param jndiName resource reference JNDI name to look up.
+     * @param metadata metadata identifier for the application artifact
+     *                     that defines the repository interface.
      * @param provider OSGi service that provides the CDI extension.
      */
-    DelegatingResourceFactory(String jndiName, String identifier, DataProvider provider) {
+    ResRefDelegator(String jndiName, String identifier, DataProvider provider) {
         this.jndiName = jndiName;
         this.identifier = identifier;
         this.provider = provider;
     }
 
+    /**
+     * Delegates to a JNDI lookup of a resource reference that was specified as the
+     * dataStore for a Repository.
+     *
+     * @param info ignored because the resource reference that was specified by the
+     *                 Repository already has its own ResourceInfo that should be
+     *                 used instead.
+     */
     @Override
     public Object createResource(ResourceInfo info) throws Exception {
         ComponentMetaData metadata = (ComponentMetaData) provider.metadataIdSvc.getMetaData(identifier);
